ref(Lotto): move to lotto pakage and handle exception.
